Add very massive clusters to last mass bin when digitizing masses
Previously, clusters with masses above log M = 15.4 were not
considered part of the largest mass bin and therefore excluded
from analysis. They are now included into the final mass bin.

For TNG-Cluster, this affects only one cluster with a mass of
log M = 15.4001, which is now added to the analysis as part of
the mass bin from 15.2 to 15.4.