Allow for file stems without a sim name
Data and figure file stems can now be set up to not contain any
mention of the simulation name. This is also now included into the
pipelines where TNG3-1 and TNG-Cluster are used at the same time
to avoid confusion (for example with data files for TNG-Cluster
having TNG300-1 in their name).