Change tracer data generated: save indices, not PIDs
Generation pipeline previously saved particle IDs which added an
extra step to using the saved data - namely matching PIDs with a
full list of PIDs. This is already done in the generation, so
repeating this step was superfluous. Pipeline now saves indices
into a contiguous array of particle data directly.