Fix bug in TNG-Cluster DAQ function
Attempting to concatenate 1D empty array with arrays of higher
dimensions lead to ValueErrors. If a field does not exist, the
KeyError is now logged for diagnostics, and nothing is appended to
the intermediate data dictionary. Empty data lists lead to a warning
log message earlier and more reliably now, too.